+++ openquake/hmtk/seismicity/max_magnitude/cumulative_moment_release.py
@@ -54,7 +54,7 @@ from openquake.hmtk.seismicity.max_magnitude.base import (
     BaseMaximumMagnitude, MAX_MAGNITUDE_METHODS)
 
 
-@MAX_MAGNITUDE_METHODS.add("get_mmax", number_bootstraps=np.int)
+@MAX_MAGNITUDE_METHODS.add("get_mmax", number_bootstraps=int)
 class CumulativeMoment(BaseMaximumMagnitude):
     '''Class to implement the bootstrapped cumulative moment estimator of
     maximum magnitude. Adapted by G. Weatherill from the Cumulative Strain
